1. UI test for spawnInApp 2. Don't restart app between UI tests.
