+++ Tests/ui-tests/UITests.swift
@@ -6,8 +6,18 @@
 
 import XCTest
 
+
 class UITests: XCTestCase {
-    private var app: XCUIApplication!
+    static var application: XCUIApplication = {
+        let app = XCUIApplication()
+        app.launch()
+        return app
+    }()
+
+    // shortcut calculated property
+    private var app: XCUIApplication {
+        return UITests.application
+    }
     
     override func setUp() {
         // Put setup code here. This method is called before the invocation of each test method in the class.
@@ -16,8 +26,6 @@ class UITests: XCTestCase {
         continueAfterFailure = false
 
         // UI tests must launch the application that they test. Doing this in setup will make sure it happens for each test method.
-        app = XCUIApplication()
-        app.launch()
         
         // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
     }
@@ -58,6 +66,19 @@ class UITests: XCTestCase {
         app.links["Click Me"].tap()
 
         let about = self.app.staticTexts["http://website/resource#something"]
+        let exists = NSPredicate(format: "exists == true")
+        let expectation1 = expectation(for: exists, evaluatedWith: about, handler: nil)
+        
+        wait(for: [expectation1], timeout: 5)
+    }
+
+    func testShowInApp2() {
+        // Tap the Left Button
+        app.buttons["Show InApp#2"].tap()
+        
+        app.links["Click Here"].tap()
+        
+        let about = self.app.staticTexts["https://www.google.com/q=something"]
         let exists = NSPredicate(format: "exists == true")
         let expectation1 = expectation(for: exists, evaluatedWith: about, handler: nil)
